Fix possible NameError in SubProcessCompiler.execute_command()
If creating the NamedTemporaryFile throws an exception (for instance if
temp_file_container doesn't exist) this triggers a NameError in the
finally block.

+++ pipeline/compilers/__init__.py
@@ -111,6 +111,7 @@ class SubProcessCompiler(CompilerBase):
             else:
                 argument_list.extend(flattening_arg)
 
+        stdout = None
         try:
             # We always catch stdout in a file, but we may not have a use for it.
             temp_file_container = cwd or os.path.dirname(stdout_captured or "") or os.getcwd()
@@ -134,8 +135,9 @@ class SubProcessCompiler(CompilerBase):
             stdout_captured = None  # Don't save erroneous result.
             raise CompilerError(e)
         finally:
-            # Decide what to do with captured stdout.
-            if stdout_captured:
-                os.rename(stdout.name, os.path.join(cwd or os.curdir, stdout_captured))
-            else:
-                os.remove(stdout.name)
+            if stdout:
+                # Decide what to do with captured stdout.
+                if stdout_captured:
+                    os.rename(stdout.name, os.path.join(cwd or os.curdir, stdout_captured))
+                else:
+                    os.remove(stdout.name)
